- refactoring (separate controller from service)  - add more fields (subject, rating)
