+++ appinfo/app.php
@@ -1,12 +1,5 @@
 <?php
 
-\OC::$server->getEventDispatcher()->addListener('OCA\Files::loadAdditionalScripts', function(){
-    \OCP\Util::addStyle('metadata', 'tabview' );
-    \OCP\Util::addScript('metadata', 'tabview' );
-    \OCP\Util::addScript('metadata', 'plugin' );
+namespace OCA\Metadata\AppInfo;
 
-    $policy = new \OCP\AppFramework\Http\EmptyContentSecurityPolicy();
-    $policy->addAllowedConnectDomain('https://nominatim.openstreetmap.org/');
-    $policy->addAllowedFrameDomain('https://www.openstreetmap.org/');
-    \OC::$server->getContentSecurityPolicyManager()->addDefaultPolicy($policy);
-});
+new Application();
